+++ src/app/[locale]/seller/reviews/[id]/page.tsx
@@ -10,7 +10,7 @@ import { OutlineBtn } from '@/components/shared/Forms/Buttons/Buttons';
 import Skeleton from '@/components/skeleton/skeleton';
 import { IReviewFeedback } from '@/constants/types';
 import { fetchReviews } from '@/services/reviewsApi'
-import { resolveDate } from '@/util/date';
+import { resolveDate } from '@/utils/date';
 import { resolveRating } from '../util/ratingUtils';
 
 import logger from '../../../../../../logger.config.mjs';
