+++ src/services/mapCenterApi.ts
@@ -1,5 +1,5 @@
 import axiosClient from "@/config/client";
-import { handleAxiosError } from "@/util/error";
+import { handleAxiosError } from "@/utils/error";
 
 import logger from '../../logger.config.mjs';
